Extending type coverage for vector
Added coverage for types byte, int8_t, int16_t, int32_t, int64_t,
                               uint8_t, uint16_t, uint32_t, uint64_t
Moved some duplicated code to common file

Not in this commit:
                type coverage for as(), convert(), get_count(), get_size()
		                  lo(), hi(), odd(), even()
                test coverage:
		  operator vector_t()
	          vec(const argTN&... args)
                  dataT()
                  convert() with different roundingMode
                  results check for convert()
                  results check for as()
                  different permutations for swizzle()
                  operator return type for swizzle() return type
                  swizzle(), XYZW_SWIZZLE, RGBA_SWIZZLE of __swizzled_vector__
                  swizzle(), XYZW_SWIZZLE, RGBA_SWIZZLE tests with non-zero value
                  XYZW_ACCESS, RGBA_ACCESS, INDEX_ACCESS
                  lo(), hi(), odd(), even() tests with non-zero value
                  load(), store() tests with non-zero value
                  load(), store() with different adressSpace
                  operator=
                  const-correctness check for operators
                  reliable control values for operators tests
                  possible better result checks for operators
                  check for return type for operators
