+++ tests/common/vector_swizzles.template
@@ -5,15 +5,15 @@
 //  Copyright:	(c) 2017 by Codeplay Software LTD. All Rights Reserved.
 //
 **************************************************************************/
-
+$IFDEF
 #define SYCL_SIMPLE_SWIZZLES
 
 #include "../common/common.h"
 #include "../common/common_vec.h"
 
-#define TEST_NAME vector_swizzles_$TYPE_NAME
+#define TEST_NAME vector_swizzles_$OPENCL$TYPE_NAME
 
-namespace vector_swizzles_$TYPE_NAME__ {
+namespace vector_swizzles_$OPENCL$TYPE_NAME__ {
 using namespace sycl_cts;
 
 /** Test each combination of vector swizzles can be generated
@@ -69,3 +69,4 @@ class TEST_NAME : public util::test_base {
 util::test_proxy<TEST_NAME> proxy;
 
 } /* namespace vector_swizzles_$TYPE_NAME__ */
+$ENDIF
